src/test/java/uk/gov/hmcts/opal/service/legacy/LegacyMiscellaneousAccountServiceTest.java
package uk.gov.hmcts.opal.service.legacy;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.slf4j.Logger;
import uk.gov.hmcts.opal.dto.search.MiscellaneousAccountSearchDto;
import uk.gov.hmcts.opal.entity.MiscellaneousAccountEntity;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertThrows;

@ExtendWith(MockitoExtension.class)
class LegacyMiscellaneousAccountServiceTest extends LegacyTestsBase {

    @Mock
    private Logger log;

    @InjectMocks
    private LegacyMiscellaneousAccountService legacyMiscellaneousAccountService;

    @Test
    void testGetMiscellaneousAccount() {
        // Arrange

        MiscellaneousAccountEntity miscellaneousAccountEntity = MiscellaneousAccountEntity.builder().build();

        // Act
        LegacyGatewayResponseException exception = assertThrows(
            LegacyGatewayResponseException.class,
            () -> legacyMiscellaneousAccountService.getMiscellaneousAccount(1)
        );

        // Assert
        assertNotNull(legacyMiscellaneousAccountService.getLog());
        assertNotNull(exception);
        assertEquals(NOT_YET_IMPLEMENTED, exception.getMessage());

    }

    @Test
    void testSearchMiscellaneousAccounts() {
        // Arrange
        MiscellaneousAccountSearchDto criteria = MiscellaneousAccountSearchDto.builder().build();

        // Act
        LegacyGatewayResponseException exception = assertThrows(
            LegacyGatewayResponseException.class,
            () -> legacyMiscellaneousAccountService.searchMiscellaneousAccounts(criteria)
        );

        // Assert
        assertNotNull(exception);
        assertEquals(NOT_YET_IMPLEMENTED, exception.getMessage());

    }

}
